1.1.6
Fixed --speed flag, it hasn't worked from version 1.1.2 and was calculating wrong in every previous version

+++ CRecv.cpp
@@ -9,7 +9,7 @@
 #include "echelonheaders.h"
 
 int clientRecv(Config cfg) {
-  int speed = cfg.speed;
+  double speed = cfg.speed;
   const char* passkey = cfg.passkey.c_str();
